# openspec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KientreAAA Web App API",
    "version": "0.1.0",
    "description": "OpenAPI spec for KientreAAA local-first Next.js web app. Covers module execution, agent runs, NotebookLM, local files, Drive helpers, settings, masked Google OAuth, and update endpoints. SSE endpoints are documented with text/event-stream responses. Secrets are never returned raw; examples use placeholders only."
  },
  "servers": [
    {
      "url": "http://localhost:3100",
      "description": "Default local production port"
    }
  ],
...
  "paths": {
...
    "/api/run": {
      "post": {
        "tags": ["run"],
        "summary": "Run hard pipeline job and stream progress over SSE",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "jobId": { "type": "string" },
                  "command": { "type": "string" },
                  "settings": { "type": "object", "additionalProperties": true }
                },
                "required": ["command"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Server-Sent Events stream",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "example": "event: log\ndata: {\"line\":\"☁️ Đang tải 1 file .docx lên Google Drive + chuyển Google Docs...\"}\n\n"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["run"],
        "summary": "Cancel a running hard pipeline job",
        "parameters": [
          { "name": "jobId", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Cancel result" } }
      }
    },
...
  },
...
}